Additional testing coverage for the nested menu
* Strange cases still exist when going directly to a url: /app/test/1.2 for example, will add a chevron.

Because the source template is non angular it is more deciding how should work, rather than re-implementing existing functionality.

These are the design decisions that needs to be made.
 * Should the menu work the same at all levels or should it work different at each level.
 * How many levels of nesting should the menu support
 * Should the chevrons appear on all leaf nodes when active

Some changes will need to be made to the admin.js file to support the decisions because the admin.js file assumes that the browser will refresh on each page change.