Service and Dto tests  * Moved UserDto and CreateUserDto tests into seperate classes  * Added DtoTestBase for validating attributes  * User dto tests just test the validation attributes, they are not reliant on the service. They run quicker than service dependant classes
Testing the dtos like this does feel like its going too far in the unit tests. There is possibly a better way to do this using attributed data on the DtoTestBase and generating tests based on the data attributes on the Dto itself. It did helped to seperate some behaviour in the service tests.