Show parent MenuItems based on child permissions
The proposed code allow parent MenuItems to be shown/hidden based on the permissions of the child items. At persent, if a user does not have access to any child MenuItems, the empty parent will always show (unless a specific permission is assigned to the parent element). With the proposed code, if the parent permissionName is 'Auto', the parent display state will be infered from the child items.